client/requirements.txt
msgpack-python==0.4.2
